stock_analysis_app/requirements.txt
streamlit
pandas
plotly
ta
requests
tabulate
python-dateutil
